refactor(admin): Q9.b — extract admin/providers.rs (GAR-470)
Extracts the "Phase 3: Providers Console" block (~340 LOC) from
`admin/handlers.rs` (3240 → 2900 LOC) into a new dedicated module
`admin/providers.rs`, following the Q9.a pattern established in PR #90.

Extracted symbols: admin_list_providers, update_provider_settings,
provider_health, enable_provider, disable_provider, provider_failover,
list_provider_overrides, set_provider_override, plus supporting request
structs. All re-exported from handlers.rs via `pub use super::providers::*`
so routes.rs callers compile unchanged.

Zero behavior change — pure mechanical move verified by 404 passing tests.

Plan: plans/0128-gar-470-admin-providers-refactor.md
Epic: GAR-430 (Quality Gates Phase 3.6)
