+++ plans/README.md
@@ -136,3 +136,4 @@ Este diretório contém os planos de implementação para o projeto GarraRUST.
 | 0125 | [GAR-623 — docs(bookkeeping): plan 0123 / PR #341 self-ref fix — CLAUDE.md + ROADMAP + plans/README](0125-gar-623-bookkeeping-plan0123-pr341-self-ref.md) | [GAR-623](https://linear.app/chatgpt25/issue/GAR-623) | ✅ Merged 2026-05-14 via PR #346 (`079b2c6`) |
 | 0126 | [`curl \| sh` auto-bootstrap PR-A: `garraia init` foundation](0126-curl-sh-auto-bootstrap-pra-wizard.md) | TBD | ✅ Merged 2026-05-14 via PR #348 (`6a2279e`) |
 | 0127 | [`curl \| sh` auto-bootstrap PR-B: `install.sh` wiring](0127-curl-sh-auto-bootstrap-prb-installer.md) | TBD | ⏳ Draft 2026-05-14 |
+| 0128 | [GAR-470 — Q9.b: extract admin/providers.rs from admin/handlers.rs (3240 → 2900 LOC)](0128-gar-470-admin-providers-refactor.md) | [GAR-470](https://linear.app/chatgpt25/issue/GAR-470) | ⏳ In Progress |
